requirements.txt
aiogram>=3.0.0
python-dotenv
jinja2
python-multipart
langchain
langchain-openai
langchain-community
psycopg2-binary
faster-whisper
aiohttp
tlgbotfwk
langsmith
pydantic
supabase>=2.0.0
postgrest>=0.10.0
requests>=2.31.0
elevenlabs>=0.2.26
openai>=1.3.0
anthropic>=0.8.0
tiktoken>=0.5.0
numpy>=1.24.0
nltk>=3.8
spacy>=3.5.0
sentence-transformers>=2.2.0
stripe>=5.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
deep-translator>=1.11.4
httpx>=0.27.0,<0.29.0